Split large builder and dev server tests
