+++ test/test_helper.exs
@@ -1,3 +1,7 @@
+Path.wildcard(Path.expand("../test_support/**/*.exs", __DIR__))
+|> Enum.sort()
+|> Enum.each(&Code.require_file/1)
+
 ExUnit.start(exclude: [:integration])
 
 Volt.Test.ExUnit.install(root: "test/volt/test/fixtures", include: ["core_api.test.ts"])
